Add permission to API routes and update the style of delete button

+++ pages/api/comment/[id].ts
@@ -1,12 +1,18 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import prisma from '../../../lib/prisma';
 import { authenticated } from '../authenticated'; 
-import { delete_comment } from '../../../utils/commentUtil';
+import { delete_comment, get_comment_by_id } from '../../../utils/commentUtil';
 
-export default authenticated(async function getCommentById(req : NextApiRequest, res : NextApiResponse ){
+export default authenticated(async function getCommentById(req : NextApiRequest, res : NextApiResponse, decoded ){
 
 	if(req.method !== 'DELETE'){
         return res.status(405).json({error : "Method not allowed, please use DELETE"});
+	}
+	
+	const comment = await get_comment_by_id(prisma, req.query.id);
+
+	if(decoded.username !== comment.author.username){
+        return res.status(401).json({error : "You have no permission on requested entity"});
     }
 
 	const deleteComment = await delete_comment(prisma, req.query.id);
